Velocity at CG fully functional (alpha needs to be tuned more). Added gimbal command from GUI. Increased IMU_TIMEOUT to remove spikes. Telemetry is no longer blocking for 10ms.

+++ main/Command.hpp
@@ -9,7 +9,7 @@ class Command {
 public:
     Command() {}
     void setup();
-    void commandGimbal(int angleX, int angleY);
+    void commandGimbal(float angleX, float angleY);
     void commandMotors(int throttleMotor1, int throttleMotor2);
     void extendLegs();
     void retractLegs();
